# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenEAAP API
  description: |
    OpenEAAP (Open Enterprise AI Agent Platform) REST API Specification.

    This API provides endpoints for managing AI agents, workflows, models, and knowledge bases.

    ## Authentication
    All API requests require authentication using Bearer tokens:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Rate Limiting
    - Default: 100 requests per minute per user
    - Burst: 200 requests per minute

    ## Versioning
    API version is included in the path: `/api/v1/...`
  version: 1.0.0
  contact:
    name: OpenEAAP Team
    email: support@openeeap.org
    url: https://github.com/openeeap/openeeap
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.openeeap.org/v1
    description: Production server
  - url: https://staging-api.openeeap.org/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /models:
    get:
      tags: [models]
      summary: List all models
      operationId: listModels
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [llm, embedding, reranker]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Model'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
      security:
        - BearerAuth: []

    post:
      tags: [models]
      summary: Register a new model
      operationId: registerModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterModelRequest'
      responses:
        '201':
          description: Model registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
      security:
        - BearerAuth: []
  /models/{modelId}:
    get:
      tags: [models]
      summary: Get model by ID
      operationId: getModel
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
      security:
        - BearerAuth: []

    delete:
      tags: [models]
      summary: Unregister model
      operationId: unregisterModel
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      responses:
        '204':
          description: Model unregistered successfully
      security:
        - BearerAuth: []
  /models/{modelId}/infer:
    post:
      tags: [models]
      summary: Run model inference
      operationId: inferModel
      parameters:
        - $ref: '#/components/parameters/ModelIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InferenceRequest'
      responses:
        '200':
          description: Inference result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferenceResponse'
      security:
        - BearerAuth: []

  # ==================== Knowledge APIs ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ModelIdParam:
      name: modelId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Model:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [llm, embedding, reranker]
        version:
          type: string
        endpoint:
          type: string
        config:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [available, unavailable]
    RegisterModelRequest:
      type: object
      required:
        - name
        - type
        - endpoint
      properties:
        name:
          type: string
        type:
          type: string
          enum: [llm, embedding, reranker]
        version:
          type: string
        endpoint:
          type: string
        config:
          type: object
    InferenceRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
        max_tokens:
          type: integer
        temperature:
          type: number
        top_p:
          type: number
        stream:
          type: boolean
    InferenceResponse:
      type: object
      properties:
        text:
          type: string
        tokens_used:
          type: integer
        finish_reason:
          type: string

    # ==================== Knowledge Schemas ====================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...